login form with authentication

+++ firebaseConfig.js
@@ -1,8 +1,10 @@
   // Import the functions you need from the SDKs you need
   import { initializeApp } from "https://www.gstatic.com/firebasejs/12.13.0/firebase-app.js";
   import { getAnalytics } from "https://www.gstatic.com/firebasejs/12.13.0/firebase-analytics.js";
+
   import { getFirestore, collection, addDoc,  doc, setDoc, getDocs, deleteDoc } from "https://www.gstatic.com/firebasejs/12.13.0/firebase-firestore.js";
-  import { getAuth,createUserWithEmailAndPassword } from "https://www.gstatic.com/firebasejs/12.13.0/firebase-auth.js";
+  
+  import { getAuth,createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut, deleteUser } from "https://www.gstatic.com/firebasejs/12.13.0/firebase-auth.js";
 
   // TODO: Add SDKs for Firebase products that you want to use
   // https://firebase.google.com/docs/web/setup#available-libraries
@@ -32,4 +34,4 @@ const db = getFirestore(app);
 const auth = getAuth(app);
 
 
-export { collection, addDoc, db, getFirestore,  doc, setDoc, getDocs, deleteDoc, auth, createUserWithEmailAndPassword }
+export { collection, addDoc, db, getFirestore,  doc, setDoc, getDocs, deleteDoc, auth, getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut, deleteUser }
